requirements.txt
aiohttp
python-decouple
psutil
tgcrypto
Telethon
